requirements.txt
boto3==1.34.149
jupyterlab==4.2.4
pandas==2.2.2
###mac###
ipython-sql==0.4.1
SQLAlchemy==1.4.52
psycopg2-binary==2.9.9
PyMySQL==1.1.1
###mac###
redshift_connector==2.1.2
sqlalchemy-redshift==0.8.14
matplotlib==3.9.1
sphinx==7.4.7
#sphinx_rtd_theme